requirements.txt
psycopg2-binary>=2.9
python-dotenv>=1.0
anthropic>=0.84
google-genai>=1.0
json-repair>=0.30
